+++ crates/pallets/starknet/src/tests/mock/setup_mock.rs
@@ -73,6 +73,7 @@ macro_rules! mock_runtime {
 				pub const ProtocolVersion: u8 = 0;
                 pub const ChainId: Felt252Wrapper = mp_chain_id::SN_GOERLI_CHAIN_ID;
                 pub const MaxRecursionDepth: u32 = 50;
+				pub const ProgramHash: Felt252Wrapper = mp_program_hash::SN_OS_PROGRAM_HASH;
             }
 
 			impl pallet_starknet::Config for MockRuntime {
@@ -88,6 +89,7 @@ macro_rules! mock_runtime {
 				type ProtocolVersion = ProtocolVersion;
                 type ChainId = ChainId;
                 type MaxRecursionDepth = MaxRecursionDepth;
+				type ProgramHash = ProgramHash;
 			}
 
 			/// Run to block n.